move all c serialization to codegen
